improve command and rename
